+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -152,7 +152,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KPQP8K7HRKSCCP3GJEWVK8DJ"),
+  _id: z.string().default("01KPQRMTMV0JJ55WRXVVDFA2ES"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,